Added a JSON Dataset loader

+++ src/main/scala/regression/io/Dataset.scala
@@ -1,5 +1,5 @@
 package regression.io
 
-class Dataset(var data: Map[Int, Int], var keysLabel: String, var valuesLabel: String) {
+class Dataset(var data: Map[Double, Double], var keysLabel: String, var valuesLabel: String) {
 
 }
